Isolate ground-state QPE tutorial tests
Our highly trained and efficient crack team of Copilots has worked to make everything just exactly perfect.

Run portable tutorial checks against every source version, retain exact snapshots for the pinned release, and narrow PySCF to cube visualization.

Feedback-by: @wavefunction91

Now everything is just exactly perfect.
